refactor(http): move ServerEndpoint to a different file
